+++ packages/react-native-sdk/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.6.21](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.6.20...@stream-io/video-react-native-sdk-0.6.21) (2024-05-07)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `1.0.1`
+* `@stream-io/video-react-bindings` updated to version `0.4.27`
 ### [0.6.20](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.6.19...@stream-io/video-react-native-sdk-0.6.20) (2024-05-07)